Fix for issue #47
The parameter $poll is not used and could be removed.
The variable $post_data does not exist. Did you forget to declare it?
It seems like you are calling the size function count() as part of the test condition (Note: Do "$i = 0, $count = count(....) in the declaration part)
Travis Error: Line 91 | ERROR | Whitespace found at end of line (Link)