Moderator view poll results
Adds a (by forum) permission for moderators to see all voters
of the polls, regardless of the poll visibility settings.
Only mods with the right permission can ask for this, and it is
only provided on demand.
There are a couple of hacks on the HTML, due to the lack of
appropriate template events (HTML div inside a h3 block, and
HTML code extending a label text from code).
Will ask for additional events, and if approved, this can be improved.

Issue #31